+++ Assets/Framework/CustomGameFramework/Runtime/Resource/Event/ResourceUpdateManifestSuccessEventArgs.cs
@@ -1,5 +1,5 @@
 /****************************************************
-*	文件：ResourceUpdateVersionAndManifestSuccessEventArgs.cs
+*	文件：ResourceUpdateManifestSuccessEventArgs.cs
 *	作者：hemingfei
 *	邮箱：hemingfei@outlook.com
 *	日期：2023/05/12 15:37:37
@@ -14,17 +14,17 @@ namespace CustomGameFramework.Runtime
 	/// <summary>
 	/// 更新版本清单成功事件
 	/// </summary>
-	public class ResourceUpdateVersionAndManifestSuccessEventArgs : GameEventArgs
+	public class ResourceUpdateManifestSuccessEventArgs : GameEventArgs
 	{
 		/// <summary>
 		/// 更新版本清单成功事件编号。
 		/// </summary>
-		public static readonly int EventId = typeof(ResourceUpdateVersionAndManifestSuccessEventArgs).GetHashCode();
+		public static readonly int EventId = typeof(ResourceUpdateManifestSuccessEventArgs).GetHashCode();
 
 		/// <summary>
 		/// 初始化更新版本清单成功事件的新实例。
 		/// </summary>
-		public ResourceUpdateVersionAndManifestSuccessEventArgs()
+		public ResourceUpdateManifestSuccessEventArgs()
 		{
 		}
 
@@ -37,9 +37,9 @@ namespace CustomGameFramework.Runtime
 		/// 创建更新版本清单成功事件。
 		/// </summary>
 		/// <returns>创建的更新版本清单成功事件。</returns>
-		public static ResourceUpdateVersionAndManifestSuccessEventArgs Create()
+		public static ResourceUpdateManifestSuccessEventArgs Create()
 		{
-			ResourceUpdateVersionAndManifestSuccessEventArgs e = ReferencePool.Acquire<ResourceUpdateVersionAndManifestSuccessEventArgs>();
+			ResourceUpdateManifestSuccessEventArgs e = ReferencePool.Acquire<ResourceUpdateManifestSuccessEventArgs>();
 			return e;
 		}