+++ Assets/Framework/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this package will be documented in this file.
 
+## [1.0.7] - 2023-05-13
+
+### Changed
+
+- Resource Download Progress Event Name Changed
+- Resource Update Manifest Evenet Name Changed
+
 ## [1.0.6] - 2023-05-12
 
 ### Added